(DONT MERGE) making all ids of type integer to safe space - having issues making database migrations

+++ src/main/java/com/zenfulcode/commercify/commercify/api/responses/PaymentResponse.java
@@ -3,9 +3,9 @@ package com.zenfulcode.commercify.commercify.api.responses;
 
 import com.zenfulcode.commercify.commercify.PaymentStatus;
 
-public record PaymentResponse(Long paymentId, PaymentStatus status, String redirectUrl) {
+public record PaymentResponse(Integer paymentId, PaymentStatus status, String redirectUrl) {
     public static PaymentResponse FailedPayment() {
-        return new PaymentResponse(-1L, PaymentStatus.FAILED, "");
+        return new PaymentResponse(-1, PaymentStatus.FAILED, "");
     }
 }
