+++ src/main/java/com/zenfulcode/commercify/commercify/repository/AddressRepository.java
@@ -3,5 +3,5 @@ package com.zenfulcode.commercify.commercify.repository;
 import com.zenfulcode.commercify.commercify.entity.AddressEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<AddressEntity, Long> {
+public interface AddressRepository extends JpaRepository<AddressEntity, Integer> {
 }
